feat: add CustomerDeliveryLineDetailFixedItems
